backend/requirements.txt
astapi==0.115.0
uvicorn==0.29.0
httpx==0.27.0
psycopg2-binary
pydantic==2.10.0
python-dotenv==1.0.1